Add batch_size to other functions except label

+++ haystack/document_store/sql.py
@@ -1,3 +1,4 @@
+import itertools
 import logging
 from typing import Any, Dict, Union, List, Optional
 from uuid import uuid4
@@ -74,7 +75,7 @@ class SQLDocumentStore(BaseDocumentStore):
         index: str = "document",
         label_index: str = "label",
         update_existing_documents: bool = False,
-        max_variable_number: int = 999,
+        batch_size: int = 999,
     ):
         """
         An SQL backed DocumentStore. Currently supports SQLite, PostgreSQL and MySQL backends.
@@ -88,9 +89,9 @@ class SQLDocumentStore(BaseDocumentStore):
                                           If set to False, an error is raised if the document ID of the document being
                                           added already exists. Using this parameter could cause performance degradation
                                           for document insertion.
-        :param max_variable_number: Maximum number of host parameters in a single SQL statement.
-                                    To help in excessive memory allocations.
-                                    More info refer: https://www.sqlite.org/limits.html
+        :param batch_size: Maximum number of host parameters in a single SQL statement.
+                           To help in excessive memory allocations.
+                           More info refer: https://www.sqlite.org/limits.html
         """
         engine = create_engine(url)
         ORMBase.metadata.create_all(engine)
@@ -101,7 +102,7 @@ class SQLDocumentStore(BaseDocumentStore):
         self.update_existing_documents = update_existing_documents
         if getattr(self, "similarity", None) is None:
             self.similarity = None
-        self.max_variable_number = max_variable_number
+        self.batch_size = batch_size
 
     def get_document_by_id(self, id: str, index: Optional[str] = None) -> Optional[Document]:
         """Fetch a document by specifying its text id string"""
@@ -112,21 +113,33 @@ class SQLDocumentStore(BaseDocumentStore):
     def get_documents_by_id(self, ids: List[str], index: Optional[str] = None) -> List[Document]:
         """Fetch documents by specifying a list of text id strings"""
         index = index or self.index
-        results = self.session.query(DocumentORM).filter(DocumentORM.id.in_(ids), DocumentORM.index == index).all()
-        documents = [self._convert_sql_row_to_document(row) for row in results]
+
+        documents = []
+        for i in range(0, len(ids), self.batch_size):
+            query = self.session.query(DocumentORM).filter(
+                DocumentORM.id.in_(ids[i: i + self.batch_size]),
+                DocumentORM.index == index
+            )
+            for row in query.all():
+                documents.append(self._convert_sql_row_to_document(row))
 
         return documents
 
     def get_documents_by_vector_ids(self, vector_ids: List[str], index: Optional[str] = None):
         """Fetch documents by specifying a list of text vector id strings"""
         index = index or self.index
-        results = self.session.query(DocumentORM).filter(
-            DocumentORM.vector_id.in_(vector_ids),
-            DocumentORM.index == index
-        ).all()
-        sorted_results = sorted(results, key=lambda doc: vector_ids.index(doc.vector_id))
-        documents = [self._convert_sql_row_to_document(row) for row in sorted_results]
-        return documents
+
+        documents = []
+        for i in range(0, len(vector_ids), self.batch_size):
+            query = self.session.query(DocumentORM).filter(
+                DocumentORM.vector_id.in_(vector_ids[i: i + self.batch_size]),
+                DocumentORM.index == index
+            )
+            for row in query.all():
+                documents.append(self._convert_sql_row_to_document(row))
+
+        sorted_documents = sorted(documents, key=lambda doc: vector_ids.index(doc.meta["vector_id"]))  # type: ignore
+        return sorted_documents
 
     def get_all_documents(
             self,
@@ -179,12 +192,12 @@ class SQLDocumentStore(BaseDocumentStore):
             if row.meta:
                 doc_ids_with_meta.append(row.id)
 
-        for i in range(0, len(doc_ids_with_meta), self.max_variable_number):
+        for i in range(0, len(doc_ids_with_meta), self.batch_size):
             meta_query = self.session.query(
                 MetaORM.document_id,
                 MetaORM.name,
                 MetaORM.value
-            ).filter(MetaORM.document_id.in_(doc_ids_with_meta[i: i + self.max_variable_number]))
+            ).filter(MetaORM.document_id.in_(doc_ids_with_meta[i: i + self.batch_size]))
 
             for row in meta_query.all():
                 documents_map[row.document_id].meta[row.name] = row.value # type: ignore
@@ -196,6 +209,7 @@ class SQLDocumentStore(BaseDocumentStore):
         Return all labels in the document store
         """
         index = index or self.label_index
+        # TODO: Use batch_size
         label_rows = self.session.query(LabelORM).filter_by(index=index).all()
         labels = [self._convert_sql_row_to_label(row) for row in label_rows]
 
@@ -216,33 +230,41 @@ class SQLDocumentStore(BaseDocumentStore):
         :return: None
         """
 
-        # Make sure we comply to Document class format
-        document_objects = [Document.from_dict(d) if isinstance(d, dict) else d for d in documents]
         index = index or self.index
-        for doc in document_objects:
-            meta_fields = doc.meta or {}
-            vector_id = meta_fields.get("vector_id")
-            meta_orms = [MetaORM(name=key, value=value) for key, value in meta_fields.items()]
-            doc_orm = DocumentORM(id=doc.id, text=doc.text, vector_id=vector_id, meta=meta_orms, index=index)
-            if self.update_existing_documents:
-                # First old meta data cleaning is required
-                self.session.query(MetaORM).filter_by(document_id=doc.id).delete()
-                self.session.merge(doc_orm)
-            else:
-                self.session.add(doc_orm)
-        try:
-            self.session.commit()
-        except Exception as ex:
-            logger.error(f"Transaction rollback: {ex.__cause__}")
-            # Rollback is important here otherwise self.session will be in inconsistent state and next call will fail
-            self.session.rollback()
-            raise ex
+        if len(documents) == 0:
+            return
+        # Make sure we comply to Document class format
+        if isinstance(documents[0], dict):
+            document_objects = [Document.from_dict(d) if isinstance(d, dict) else d for d in documents]
+        else:
+            document_objects = documents
+
+        for i in range(0, len(document_objects), self.batch_size):
+            for doc in document_objects[i: i + self.batch_size]:
+                meta_fields = doc.meta or {}
+                vector_id = meta_fields.get("vector_id")
+                meta_orms = [MetaORM(name=key, value=value) for key, value in meta_fields.items()]
+                doc_orm = DocumentORM(id=doc.id, text=doc.text, vector_id=vector_id, meta=meta_orms, index=index)
+                if self.update_existing_documents:
+                    # First old meta data cleaning is required
+                    self.session.query(MetaORM).filter_by(document_id=doc.id).delete()
+                    self.session.merge(doc_orm)
+                else:
+                    self.session.add(doc_orm)
+            try:
+                self.session.commit()
+            except Exception as ex:
+                logger.error(f"Transaction rollback: {ex.__cause__}")
+                # Rollback is important here otherwise self.session will be in inconsistent state and next call will fail
+                self.session.rollback()
+                raise ex
 
     def write_labels(self, labels, index=None):
         """Write annotation labels into document store."""
 
         labels = [Label.from_dict(l) if isinstance(l, dict) else l for l in labels]
         index = index or self.label_index
+        # TODO: Use batch_size
         for label in labels:
             label_orm = LabelORM(
                 document_id=label.document_id,
@@ -267,16 +289,22 @@ class SQLDocumentStore(BaseDocumentStore):
         :param index: filter documents by the optional index attribute for documents in database.
         """
         index = index or self.index
-        self.session.query(DocumentORM).filter(
-            DocumentORM.id.in_(vector_id_map),
-            DocumentORM.index == index
-        ).update({
-            DocumentORM.vector_id: case(
-                vector_id_map,
-                value=DocumentORM.id,
-            )
-        }, synchronize_session=False)
-        self.session.commit()
+        for chunk_map in self.chunked_iterable(vector_id_map.items(), size=self.batch_size):
+            self.session.query(DocumentORM).filter(
+                DocumentORM.id.in_(chunk_map),
+                DocumentORM.index == index
+            ).update({
+                DocumentORM.vector_id: case(
+                    chunk_map,
+                    value=DocumentORM.id,
+                )
+            }, synchronize_session=False)
+            try:
+                self.session.commit()
+            except Exception as ex:
+                logger.error(f"Transaction rollback: {ex.__cause__}")
+                self.session.rollback()
+                raise ex
 
     def update_document_meta(self, id: str, meta: Dict[str, str]):
         """
@@ -385,3 +413,12 @@ class SQLDocumentStore(BaseDocumentStore):
             session.add(instance)
             session.commit()
             return instance
+
+    # Refer https://alexwlchan.net/2018/12/iterating-in-fixed-size-chunks/
+    def chunked_iterable(self, iterable, size):
+        it = iter(iterable)
+        while True:
+            chunk = tuple(itertools.islice(it, size))
+            if not chunk:
+                break
+            yield chunk
